+++ tomcat/src/main/java/org/apache/catalina/http11/Dispatcher.java
@@ -6,7 +6,7 @@ import org.apache.catalina.controller.HomeController;
 import org.apache.catalina.controller.LoginController;
 import org.apache.catalina.controller.RegisterController;
 import org.apache.catalina.controller.StaticController;
-import org.apache.catalina.http.HeaderName;
+import org.apache.catalina.HeaderName;
 import org.apache.catalina.request.HttpRequest;
 import org.apache.catalina.response.HttpResponse;
