+++ tomcat/src/main/java/org/apache/coyote/connector/RequestReader.java
@@ -1,4 +1,4 @@
-package org.apache.catalina.request;
+package org.apache.coyote.connector;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -6,6 +6,7 @@ import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.util.ArrayList;
 import java.util.List;
+import org.apache.catalina.request.HttpRequest;
 
 public class RequestReader {
